update exercises and solutions for testing

+++ testing/solutions/counting-tests/test_counting.py
@@ -3,11 +3,11 @@ import numpy as np
 from numpy.testing import assert_equal
 
 def test_count_no_object():
-    a = np.zeros([6, 6], dtype=np.float64)
+    a = np.zeros([6, 6], dtype=np.int32)
     assert_equal(count_specimens(a), 0)
 
 def test_count_1_object():
-    a = np.ones([6, 6], dtype=np.float64)
+    a = np.ones([6, 6], dtype=np.int32)
     assert_equal(count_specimens(a), 1)
 
 def test_count_specimens_2_objects():
@@ -19,7 +19,7 @@ def test_count_specimens_2_objects():
        [ 0.,  0.,  0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  1.,  1.,  0.],
        [ 0.,  0.,  0.,  0.,  1.,  1.,  0.],
-       [ 0.,  0.,  0.,  0.,  0.,  0.,  0.]])
+       [ 0.,  0.,  0.,  0.,  0.,  0.,  0.]], dtype=np.int32)
     assert_equal(count_specimens(a), 2)
 
 def test_count_1_object_rectangle():
@@ -33,5 +33,5 @@ def test_count_1_object_rectangle():
        [ 0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  0.],
        [ 0.,  0.,  0.,  0.,  0.],
-       [ 0.,  0.,  0.,  0.,  0.]])
+       [ 0.,  0.,  0.,  0.,  0.]], dtype=np.int32)
     assert_equal(count_specimens(a), 1)
